Split triggers.cs into Start/EndTouch
