+++ src/ST-Events/Tick.cs
@@ -8,10 +8,6 @@ public partial class SurfTimer
         {
             player.Timer.Tick();
             player.HUD.Display();
-
-            #if DEBUG
-            if (player.Controller.IsValid && player.Controller.PawnIsAlive) player.Controller.PrintToCenter($"DEBUG >> PrintToCenter -> Player.Timer.Ticks: {player.Timer.Ticks}");
-            #endif
         }
     }
 }
